% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCcontent.R
\name{getGCpercentage}
\alias{getGCpercentage}
\title{calulate GC for several window sizes}
\usage{
getGCpercentage(sites, column_prefix, window_size, reference_genome_sequence)
}
\arguments{
\item{sites}{GRange locations on genome: chromosome, strand, position}

\item{column_prefix}{what prefix to add to each column}

\item{window_size}{vector with names: elements are window, names used as col names}

\item{reference_genome_sequence}{actual reference genome (from BSgenome.* object)
to get reference_genome_sequence use get_reference_genome() function from intSiteRetiever}
}
\value{
Grange object with added columns: prefix.name_window_size 
and value equal to GC content for a window
}
\description{
calulate GC for several window sizes
}
\note{
start up time to load human genome is about 10 second
}
